requirements.txt
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
fonttools==4.53.1
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gpxpy==1.6.2
h3==3.7.7
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.1
moviepy==1.0.3
numpy==2.0.1
odfpy==1.4.1
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
Rtree==1.3.0
setuptools==72.1.0
shapely==2.0.6
six==1.16.0
timezonefinder==6.5.2
tqdm==4.66.4
urllib3==2.2.2
